requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
reportlab==4.0.7
python-docx==1.1.0
python-dateutil==2.8.2
Werkzeug==2.3.7
